Create EggDropCrane
This code was used with an arduino based egg dropping robot. The robot hangs from the ceiling by a strong fishing line. It uses a hoist to raise and lower itself to pre-defined heights. Distance measurements are obtained by two sonar sensors so robot knows distance from floor and distance from ceiling. A servo controls release of egg basket which drops an egg. Commands are issued to the robot via bluetooth serial connection. I used BlueTerm on android to interface with robot.